+++ sources/js/game/controller.js
