+++ sources/js/game/model.js
