+++ sources/js/game/views/game.js
