+++ sources/js/game/views/lifes.js
