+++ sources/js/game/views/pause/template.tmpl
