+++ sources/js/game/views/ready/template.tmpl
