+++ sources/js/game/views/score.js
