+++ sources/js/game/views/start-menu/template.tmpl
